Travis: fix the GCC+boost::regex builds by updating the compiler versions
The old Travis file had gcc-4.8 + boost::regex. But I think the default
Ubuntu install (and the pre-packaged libboost_regex.so) was compiled against
libstdc++-5 which had a major ABI break with std::string, which caused an ABI
problem when gcc-4.8 was linking aginst that library.

Easiest thing is to stop testing gcc-4.8
Also updated the clang and gcc versions to try newer ones.