Check how docopt is used to decide whether generate install target or not.
When I try to use docopt through `FecthContent` of cmake, docopt.cpp
generate the install target as well, which is annoying.

Although there is [a way][stackoverflow-link] to exclude the
installation of docopt.cpp, it's not so elegant.

I write this patch by checking how spdlog handling this not-root-project
situation. It seems work.